+++ NEWS.md
@@ -24,6 +24,7 @@
 
 # ggtree 4.1.1.002
 
++ more options in `ggtree_set_interactive()` (2025-12-19, Fri)
 + update `default_aes` to work with `from_theme`
   - `geom_highlight` and `geom_taxalink`  (2025-12-18, Thu, #694)
 + fix issue of `geom_taxalink` by adding 'outward' argument (2025-12-18, Thu, #692)